chore: 파일 위치 변경
- 쿼리 함수 model로 이동
- API 함수 memberSearchApi로 흡수
- generationListQueryOptions 파일 queryOptions 파일로 흡수
