+++ apps/web/src/features/member-search/api/useMemberSearch.ts
@@ -1,5 +1,5 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
-import { memberSearchQueryOptions } from '@/features/member-search/api/queryOptions';
+import { memberSearchQueryOptions } from '@/features/member-search/model/queries/queryOptions';
 import { MemberSearchFilters } from '@/entities/search/model/types';
 
 export function useMemberSearch(filters: MemberSearchFilters) {
