+++ apps/web/src/features/member-search/model/queries/queryOptions.ts
@@ -1,7 +1,7 @@
-import { infiniteQueryOptions } from '@tanstack/react-query';
+import { infiniteQueryOptions, queryOptions } from '@tanstack/react-query';
 import { memberSearchQueryKeys } from './queryKeys';
 import { memberSearchApi } from '@/features/member-search/api/memberSearchApi';
-import { toMemberSearchRequest, mapMemberSearchItem } from '../model/mapper';
+import { toMemberSearchRequest, mapMemberSearchItem, toGenerationList } from '../mapper';
 import { MemberSearchFilters } from '@/entities/search/model/types';
 import { MemberSearchResponse } from '@/features/member-search/api/types';
 
@@ -40,6 +40,21 @@ export function memberSearchQueryOptions(filters: MemberSearchFilters) {
         totalCount: data.pages[0]?.totalCount ?? null,
         isLast: data.pages[data.pages.length - 1]?.isLast ?? true,
       };
+    },
+  });
+}
+
+/**
+ * 기수 목록 조회 Query 옵션
+ *
+ * queryFn에서 DTO -> 도메인 모델 매핑을 수행해 캐시에 저장합니다.
+ */
+export function generationListQueryOptions() {
+  return queryOptions({
+    queryKey: memberSearchQueryKeys.generations(),
+    queryFn: async () => {
+      const dto = await memberSearchApi.getGenerationList();
+      return toGenerationList(dto);
     },
   });
 }
